# openapi/api.yaml
openapi: 3.0.3
info:
  title: RevU API
  description: RevU API for businesses and customers.
  version: 1.0.2
servers:
  - url: 'http://34.209.63.8:5000/api/v1'
    description: Production
  - url: 'http://localhost:5000/api/v1'
    description: Development
security:
  - bearerAuth: []
paths:
...
  /services/categories:
    post:
      summary: Create a serviceCategory
      tags:
        - serviceCategories
      operationId: createServiceCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceCategory'
            examples:
              serviceCategory:
                $ref: '#/components/examples/serviceCategory'
      responses:
        '201':
          description: ServiceCategory created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ServiceCategory'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    get:
      summary: Get serviceCategories list
      tags:
        - serviceCategories
      operationId: getServiceCategories
      parameters:
        - $ref: '#/components/parameters/parentId'
        - $ref: '#/components/parameters/applicableSex'
        - $ref: '#/components/parameters/salonIds'
        - $ref: '#/components/parameters/servicesIds'
        - $ref: '#/components/parameters/serviceCategoryIncludeWithPagination'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/perPage'
      responses:
        '200':
          description: Paginated serviceCategories list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedListSuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ServiceCategory'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    delete:
      summary: Delete serviceCategories
      tags:
        - serviceCategories
      operationId: deleteServiceCategories
      parameters:
        - $ref: '#/components/parameters/deleteIds'
      responses:
        '200':
          description: ServiceCategories deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiObjectOperationSuccessResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /services/categories/{id}:
    get:
      summary: Get a serviceCategory
      tags:
        - serviceCategories
      operationId: getServiceCategory
      parameters:
        - $ref: '#/components/parameters/id'
        - $ref: '#/components/parameters/childrenInclude'
      responses:
        '200':
          description: ServiceCategory details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ServiceCategory'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /services:
    post:
      summary: Create a service
      tags:
        - services
      operationId: createService
      parameters:
        - $ref: '#/components/parameters/serviceInclude'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Service'
            examples:
              service:
                $ref: '#/components/examples/service'
      responses:
        '201':
          description: Service created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Service'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    get:
      summary: Get services list
      tags:
        - services
      operationId: getServices
      parameters:
        - $ref: '#/components/parameters/applicableSex'
        - $ref: '#/components/parameters/categoryIds'
        - $ref: '#/components/parameters/serviceType'
        - $ref: '#/components/parameters/serviceIncludeWithPagination'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/perPage'
      responses:
        '200':
          description: Paginated services list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedListSuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Service'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    delete:
      summary: Delete services
      tags:
        - services
      operationId: deleteServices
      parameters:
        - $ref: '#/components/parameters/deleteIds'
      responses:
        '200':
          description: Services deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiObjectOperationSuccessResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /services/{id}:
    get:
      summary: Get a service
      tags:
        - services
      operationId: getService
      parameters:
        - $ref: '#/components/parameters/id'
        - $ref: '#/components/parameters/serviceInclude'
      responses:
        '200':
          description: Service details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Service'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
...
  /timeSlots:
    post:
      summary: Create a timeSlot
      tags:
        - timeSlots
      operationId: createTimeSlot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeSlot'
            examples:
              timeSlot:
                $ref: '#/components/examples/timeSlot'
      responses:
        '201':
          description: TimeSlot created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TimeSlot'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    get:
      summary: Get timeSlots list
      tags:
        - timeSlots
      operationId: getTimeSlots
      parameters:
        - $ref: '#/components/parameters/salonId'
        - $ref: '#/components/parameters/timeSlotAvailabilities'
        - $ref: '#/components/parameters/optedServicesIds'
        - $ref: '#/components/parameters/employeeIds'
        - $ref: '#/components/parameters/paginationInclude'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/perPage'
      responses:
        '200':
          description: Paginated timeSlots list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedListSuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/TimeSlot'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    delete:
      summary: Delete timeSlots
      tags:
        - timeSlots
      operationId: deleteTimeSlots
      parameters:
        - $ref: '#/components/parameters/deleteIds'
      responses:
        '200':
          description: TimeSlots deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiObjectOperationSuccessResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /timeSlots/{id}:
    get:
      summary: Get a timeSlot
      tags:
        - timeSlots
      operationId: getTimeSlot
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: TimeSlot details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TimeSlot'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
...
components:
  parameters:
    id:
      name: id
      description: Resource id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/IntegerId'
...
    parentId:
      name: parentId
      description: Filter by resource parent id. Value `0` is interpreted as `null`.
      in: query
      schema:
        $ref: '#/components/schemas/PositiveIntegerIncludingZero'
    deleteIds:
      name: ids
      description: List of ids to delete
      in: query
      required: true
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/IntegerIds'
    categoryIds:
      name: categoryIds
      description: Limit result to matching category ids
      in: query
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/IntegerIds'
...
    salonId:
      name: salonId
      description: Limit results to specific salon id
      in: query
      schema:
        $ref: '#/components/schemas/IntegerId'
...
    employeeIds:
      name: employeeIds
      description: Limit result to matching employee ids
      in: query
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/IntegerIds'
...
    salonIds:
      name: salonIds
      description: Limit result to matching salon ids
      in: query
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/IntegerIds'
    servicesIds:
      name: servicesIds
      description: Limit result to matching service ids
      in: query
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/IntegerIds'
    optedServicesIds:
      name: optedServicesIds
      description: Limit result to matching optedServices ids
      in: query
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/IntegerIds'
    applicableSex:
      name: applicableSex
      description: Limit results relating to a specific sex
      in: query
      schema:
        $ref: '#/components/schemas/ApplicableSex'
    serviceType:
      name: type
      description: Limit results to specific service type
      in: query
      schema:
        $ref: '#/components/schemas/ServiceType'
...
    timeSlotAvailabilities:
      name: availabilities
      description: Limit results to with matching availabilities
      in: query
      style: form
      explode: false
      schema:
        type: array
        items:
          $ref: '#/components/schemas/TimeSlotAvailability'
...
    paginationInclude:
      name: include
      description: Get extra fields in result
      in: query
      style: form
      explode: false
      schema:
        type: array
        items:
          enum:
          - _pagination.total
...
    serviceInclude:
      name: include
      description: Get extra fields in result
      in: query
      style: form
      explode: false
      schema:
        type: array
        items:
          enum:
            - categories
    serviceIncludeWithPagination:
      name: include
      description: Get extra fields in result
      in: query
      style: form
      explode: false
      schema:
        type: array
        items:
          enum:
            - categories
            - image
            - _pagination.total
...
    childrenInclude:
      name: include
      description: Get extra fields in result
      in: query
      style: form
      explode: false
      schema:
        type: array
        items:
          enum:
            - children
    serviceCategoryIncludeWithPagination:
      name: include
      description: Get extra fields in result
      in: query
      style: form
      explode: false
      schema:
        type: array
        items:
          enum:
            - children
            - image
            - _pagination.total
...
    page:
      name: page
      description: Current page no. for pagination
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    perPage:
      name: perPage
      description: No. of records in paginated list
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
...
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    '500':
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    PositiveIntegerIncludingZero:
      type: integer
      minimum: 0
    IntegerId:
      type: integer
      minimum: 1
    IntegerIds:
      type: array
      items:
        $ref: '#/components/schemas/IntegerId'
      minItems: 1
...
    MediumText:
      type: string
      minLength: 2
      maxLength: 255
...
    Money:
      type: number
      format: float
      minimum: 0
...
    DateTime:
      type: string
      format: date-time
      description: ISO 8601 compliant date time
    FileType:
      enum:
        - image
    MimeType:
      enum:
        - image/jpeg
        - image/png
        - image/gif
        - image/webp
        - image/svg+xml
...
    ApplicableSex:
      enum:
        - male
        - female
        - other
        - unisex
        - all
...
    ServiceStyleType:
      enum:
        - trending
        - retro
        - modern
        - indian
        - western
...
    TimeSlotAvailability:
      enum:
        - available
        - occupied
        - temporaryUnavailable
        - unavailable
...
    ServiceType:
      enum:
        - standard
        - additional
...
    File:
      type: object
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/IntegerId'
          readOnly: true
        name:
          allOf:
            - $ref: '#/components/schemas/MediumText'
          readOnly: true
        type:
          allOf:
            - $ref: '#/components/schemas/FileType'
          readOnly: true
        title:
          $ref: '#/components/schemas/MediumText'
        mimeType:
          allOf:
            - $ref: '#/components/schemas/MimeType'
          readOnly: true
        url:
          type: string
          format: uri
          readOnly: true
        thumbnailURL:
          type: string
          format: uri
          readOnly: true
        updatedAt:
          $ref: '#/components/schemas/DateTime'
      required:
        - title
        - type
        - url
...
    Service:
      type: object
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/IntegerId'
          readOnly: true
        name:
          $ref: '#/components/schemas/MediumText'
        type:
          $ref: '#/components/schemas/ServiceType'
        applicableSex:
          $ref: '#/components/schemas/ApplicableSex'
        description:
          type: string
        defaultPrice:
          $ref: '#/components/schemas/Money'
        defaultDurationSeconds:
          type: integer
          minimum: 60
        image:
          allOf:
            - $ref: '#/components/schemas/File'
          readOnly: true
        imageId:
          $ref: '#/components/schemas/IntegerId'
        styles:
          type: array
          items:
            $ref: '#/components/schemas/ServiceStyleType'
        categoryIds:
          allOf:
            - $ref: '#/components/schemas/IntegerIds'
            - writeOnly: true
      required:
        - name
        - type
        - applicableSex
...
    ServiceCategory:
      type: object
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/IntegerId'
          readOnly: true
        name:
          $ref: '#/components/schemas/MediumText'
        applicableSex:
          $ref: '#/components/schemas/ApplicableSex'
        image:
          allOf:
            - $ref: '#/components/schemas/File'
          readOnly: true
        imageId:
          $ref: '#/components/schemas/IntegerId'
        bgColor:
          type: string
          format: regex
          pattern: '^#([0-9a-fA-F]{3}){1,2}$'
        parent:
          allOf:
            - $ref: '#/components/schemas/ServiceCategory'
          readOnly: true
        parentId:
          $ref: '#/components/schemas/IntegerId'
        children:
          type: array
          items:
            $ref: '#/components/schemas/ServiceCategory'
          readOnly: true
        updatedAt:
          allOf:
            - $ref: '#/components/schemas/DateTime'
          readOnly: true
      required:
        - name
        - applicableSex
        - bgColor
    TimeSlot:
      type: object
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/IntegerId'
            - readOnly: true
        employeeId:
          $ref: '#/components/schemas/IntegerId'
        start:
          $ref: '#/components/schemas/DateTime'
        end:
          $ref: '#/components/schemas/DateTime'
        availability:
          $ref: '#/components/schemas/TimeSlotAvailability'
        note:
          type: string
      required:
        - employeeId
        - start
        - end
        - availability
...
    Error:
      type: object
      properties:
        code:
          oneOf:
            - type: number
            - type: string
        message:
          type: string
        data:
          oneOf:
            - type: object
            - type: array
              items: {}
      required:
        - message
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        perPage:
          type: integer
          minimum: 1
        total:
          type: integer
      required:
        - page
        - perPage
    JSONResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/JSONResponse'
        - type: object
          properties:
            success:
              type: boolean
              default: true
    SingleSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
          required:
            - data
    ListSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items: {}
          required:
            - data
    PaginatedListSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/ListSuccessResponse'
        - type: object
          properties:
            _pagination:
              $ref: '#/components/schemas/Pagination'
          required:
            - _pagination
...
    MultiObjectOperationSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          default: true
        data:
          type: object
          properties:
            count:
              $ref: '#/components/schemas/PositiveIntegerIncludingZero'
          required:
            - count
      required:
        - success
        - data
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/JSONResponse'
        - type: object
          properties:
            error:
              $ref: '#/components/schemas/Error'
          required:
            - error
  examples:
...
    service:
      value:
        name: Hair Cut
        type: standard
        applicableSex: unisex
        description: This is a best hair cut service.
        defaultPrice: 30
        defaultDurationSeconds: 3600
        imageId: 1
        styles: [trending, modern]
        categoryIds: [1, 3]
    serviceCategory:
      value:
        name: Hair Care
        applicableSex: unisex
        imageId: 1
        bgColor: '#123456'
...
    timeSlot:
      value:
        employeeId: 1
        start: '2022-08-15T13:26:34.769Z'
        end: '2022-08-15T19:26:34.769Z'
        availability: available
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT